api/requirements.txt
fastapi
uvicorn
valkey
pydantic
pwdlib
pyjwt
python-dotenv
